refactor localization module into separate modules
module was bloated, now have folder with __init__ that imports the public-facing API classes/functions, and separate modules for localization_algorithms.py, spatial_event.py, synchronized_recorder_array.py

refactored top-level imports and tests accordingly
